Create Milestone 1
You must program one in the first milestone web server with programming language C. 
Functional requirements:

1. The server must provide "as-is" files. 
As-is" files are files that contain both http header and http body, and sent "as is", without adding or subtracting anything. It does not get any easier: the server opens the file and writes it to to the socket connected to the client. These files should end up .asisand shall contain both http-head and http-body , as in the example below. 
./eksempler/index.asis 

HTTP/1.1 200 OK
Content-Type: text/plain
Content-Length: 16
Hello World :-)

2. If there is an http request for a file that does not exist , or to a file of an unsupported type , then correct http error message returned to the client.
3. Since the server must be demonized (see Security / Robustness Requirements below), then it has no control terminal to print to. 

Safety / robustness requirements

1. The program should not commit to the gate once per. request, but only at program start.
2. Each client request must be processed in a separate thread or process.
3. The server must be demonized so that it is independent of the control terminal.
4. The server it listens on port 80 without it running as the user root .
5. The directory that contains the files that should be accessible via http should be on web root . The server processes should too (using chroot (2) ) have changed their root directory to this one this web site ..