Create Milestone 2
1. Extend the server from Milestone 1 so that it (in addition to asis files) can provide all file types listed in fly /etc/mime.types. The identification of type must be done using the file extension. If e.g. the file ends on .txt so be Content-type set to text/plain.

2. Create a web page for the group as:

       a) has at least one image,
       b) is be styled with a CSS file and
       c) has a hyperlink to an XML file that displays info about the group members

All files (HTML, CSS, XML and image files) as the web page consists of must be stored in the webbot, or a subdirectory of this. 

3. Make sure the server is running in a busybox-based container created using chroot and unshare, allowing access to files and processes are limited. Feel free to use the example as a starting point unshare-container I or unshare-container II. It shall still also be possible to run the server as a demon directly on the host system - ie outside the container.

The felt tree in the container must contain as little as possible. It can see, for example something like this: 


Name suggestions 	           comment

/bin/ 	                          Here are the busybox links and the web server
/proc/ 	                          A virtual file system for process information
/var/www/ 	                  A directory that should be the web site.
/var/log/debug.log 	          Log file to be connected to STDERR. This should not be in the web root.
/etc/mime.types 	          To identify mime type by file last name

To run the web server in a container, it is a good idea to have one init process running as process 1. The web services can then be run as a child of init (eg with PID = 2). 
Some of the safety / robustness requirements from Milestone 1 are unnecessary, when the server is run in the container. You can therefore enter a test, so that unnecessary safety / robustness measures are not performed if the parenting process is init. For example. such: 

if  (1!=getppid()) {
  tiltak(); 
}

This is how the process will carry out the measure if it is started normally from the command line outside the container. Inside the container will shell the measure has already been taken.